refactored enigma and added modules for encrypt decrypt key_generator and today_date
