+++ lib/decrypt_runner.rb
@@ -1,0 +1,1 @@
+require 'enigma'
