+++ lib/today_date.rb
@@ -1,0 +1,5 @@
+module TodayDate
+  def today_date
+    date = Date.today.strftime("%m%d%y")
+  end
+end
